+++ src/app/types.ts
@@ -1,0 +1,11 @@
+export type GeneratedCode = {
+  filePath: string;
+  fileContent: string;
+  fileType: "api-ts" | "functions-ts";
+};
+
+export type GenerateCodeInput = {
+  openApiUri: string;
+  headers?: string;
+  baseUrl?: string;
+};